Fix train from file by moving argument handling to lib/utils
